src/train/torchtask/requirements.txt
numpy
scipy
Pillow
pyyaml
opencv-python
